Split `intervals::ops::set_ops`
