+++ src/intervals/ops/set_ops/intersect_tests.rs
@@ -1,0 +1,291 @@
+use std::error::Error;
+
+use jiff::Zoned;
+
+use super::intersect::*;
+use crate::intervals::absolute::{
+    AbsoluteBoundPair,
+    AbsoluteEndBound,
+    AbsoluteFiniteBound,
+    AbsoluteStartBound,
+    EmptiableAbsoluteBoundPair,
+};
+use crate::intervals::meta::BoundInclusivity;
+use crate::ops::IntersectionResult;
+
+#[test]
+fn empty_empty() {
+    assert_eq!(
+        EmptiableAbsoluteBoundPair::Empty.intersect(&EmptiableAbsoluteBoundPair::Empty),
+        IntersectionResult::Separate,
+    );
+}
+
+#[test]
+fn empty_unbounded() {
+    assert_eq!(
+        EmptiableAbsoluteBoundPair::Empty.intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::InfinitePast,
+            AbsoluteEndBound::InfiniteFuture,
+        )),
+        IntersectionResult::Separate,
+    );
+}
+
+#[test]
+fn unbounded_empty() {
+    assert_eq!(
+        AbsoluteBoundPair::new(AbsoluteStartBound::InfinitePast, AbsoluteEndBound::InfiniteFuture,)
+            .intersect(&EmptiableAbsoluteBoundPair::Empty),
+        IntersectionResult::Separate,
+    );
+}
+
+#[test]
+fn unbounded_unbounded() {
+    assert_eq!(
+        AbsoluteBoundPair::new(AbsoluteStartBound::InfinitePast, AbsoluteEndBound::InfiniteFuture,).intersect(
+            &AbsoluteBoundPair::new(AbsoluteStartBound::InfinitePast, AbsoluteEndBound::InfiniteFuture,)
+        ),
+        IntersectionResult::Intersected(AbsoluteBoundPair::new(
+            AbsoluteStartBound::InfinitePast,
+            AbsoluteEndBound::InfiniteFuture,
+        )),
+    );
+}
+
+#[test]
+fn bounded_no_overlap() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-04 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )),
+        IntersectionResult::Separate,
+    );
+
+    Ok(())
+}
+
+#[test]
+fn bounded_adjacent_inclusive_inclusive() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )),
+        IntersectionResult::Intersected(AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )),
+    );
+
+    Ok(())
+}
+
+#[test]
+fn bounded_adjacent_inclusive_exclusive() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Exclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )),
+        IntersectionResult::Separate,
+    );
+
+    Ok(())
+}
+
+#[test]
+fn bounded_adjacent_exclusive_inclusive() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Exclusive,
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )),
+        IntersectionResult::Separate,
+    );
+
+    Ok(())
+}
+
+#[test]
+fn bounded_adjacent_exclusive_exclusive() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Exclusive,
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Exclusive,
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new_with_inclusivity(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp(),
+                BoundInclusivity::Inclusive,
+            )),
+        )),
+        IntersectionResult::Separate,
+    );
+
+    Ok(())
+}
+
+#[test]
+fn bounded_clear_overlap() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-04 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )),
+        IntersectionResult::Intersected(AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-03 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )),
+    );
+
+    Ok(())
+}
+
+#[test]
+fn bounded_on_unbounded() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )
+        .intersect(&AbsoluteBoundPair::new(
+            AbsoluteStartBound::InfinitePast,
+            AbsoluteEndBound::InfiniteFuture,
+        )),
+        IntersectionResult::Intersected(AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )),
+    );
+
+    Ok(())
+}
+
+#[test]
+fn unbounded_on_bounded() -> Result<(), Box<dyn Error>> {
+    assert_eq!(
+        AbsoluteBoundPair::new(AbsoluteStartBound::InfinitePast, AbsoluteEndBound::InfiniteFuture,).intersect(
+            &AbsoluteBoundPair::new(
+                AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                    "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+                )),
+                AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                    "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+                )),
+            )
+        ),
+        IntersectionResult::Intersected(AbsoluteBoundPair::new(
+            AbsoluteStartBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-01 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+            AbsoluteEndBound::Finite(AbsoluteFiniteBound::new(
+                "2025-01-02 00:00:00[Europe/Oslo]".parse::<Zoned>()?.timestamp()
+            )),
+        )),
+    );
+
+    Ok(())
+}
